requirements.txt
pandas==1.5.2
transformers==4.25.1
Pillow==9.3.0
torch==1.13.1